피드백 받은 부분 적용 클래스 분리
