+++ src/main/java/caculator/Caculator.java
@@ -4,10 +4,8 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class Caculator {
-    private static final int NULL_EMPTY_VALUE = 0;
-    private static final int NORMAL_VALUE = 1;
+
     private static final String SPLIT_TEXT = ",|:";
-    private static final String MINUS_OPERATOR = "-";
     private static final String CUSTOM_SPLIT_REGEX = "//(.)\n(.*)";
 
     private String inputText;
@@ -17,14 +15,6 @@ public class Caculator {
 
     public Caculator(String inputText) {
         this.inputText = inputText;
-    }
-
-    public int validateInputText(String inputText) {
-        if (inputText == null || inputText.isEmpty()) {
-            return NULL_EMPTY_VALUE;
-        }
-
-        return NORMAL_VALUE;
     }
 
     private int oneNumberInputText(String inputText) {
@@ -57,10 +47,4 @@ public class Caculator {
 
     }
 
-    public void inputTextIsNegative(String inputText) {
-        if (inputText.contains(MINUS_OPERATOR)) {
-            throw new RuntimeException("음수가 들어 올 수 없습니다.");
-        }
-
-    }
 }
